Swap PGlite for wa-sqlite as the local storage engine (Phase 2)
Completes the migration started in the previous commit: the local
database now runs on wa-sqlite (SQLite-in-WASM) instead of PGlite
(Postgres-in-WASM), cutting the local engine's payload from ~13.5MB to
~1.14MB and fixing the underlying cause of the 10+ second Stock-screen
load. Remote Postgres/Supabase is untouched.

wa-sqlite has no built-in Worker/RPC wrapper the way PGlite did, so
worker.ts and client.ts hand-roll the small piece that convenience used
to provide: a postMessage request/response protocol with one pending
promise per in-flight call. Uses the IDBBatchAtomicVFS example VFS for
IndexedDB-backed persistence (see the engine-choice reasoning in the
migration plan/memory).

Two real bugs only surfaced by actually running this in a browser
(tsc's cached incremental build was silently not re-checking wa-sqlite
usage at all -- a full `tsc -b` from a clean cache does catch it):

- src/lib/sync.ts and src/lib/weather.ts each had one leftover
  Postgres-only query (a `::int` cast, a bare `now()`) that the
  previous commit's queries.ts/syncCore.ts audit missed entirely.
- wa-sqlite's async build runs on Asyncify, which is not reentrant --
  two overlapping calls into the WASM module corrupt it outright
  ("memory access out of bounds", "Aborted(RuntimeError: unreachable)"),
  which is exactly what a few effects each firing their own query on
  mount does. Fixed by serializing every worker RPC call onto a single
  chained queue.

Verified end-to-end in a real browser (Playwright): boots cleanly with
zero console errors, lists/adds records, and data survives a full page
reload (IndexedDB persistence). Production build's actual output
confirms the payload win: wa-sqlite-async.wasm is 1,139KB against
PGlite's previous 8,090KB + 5,402KB, and the worker bundle dropped from
418KB to 85KB.

Not yet done: the cutover gate for devices with existing PGlite data
(next phase), and removing @electric-sql/pglite from package.json
(only once the cutover ships and is confirmed working for existing
users -- it's still a live dependency until then, unused by app code
but not yet safe to delete).

+++ src/lib/sync.ts
@@ -85,7 +85,7 @@ async function run(): Promise<SyncResult> {
 
 export async function pendingCount(): Promise<number> {
   const { rows } = await local.query<{ n: number }>(
-    `select count(*)::int as n from sync_outbox`,
+    `select count(*) as n from sync_outbox`,
   )
   return rows[0]?.n ?? 0
 }
